ref: move find_dicoms out of utils module
