+++ src/imgtools/dicom/__init__.py
@@ -1,4 +1,5 @@
 from .dicom_metadata import all_modalities_metadata, get_modality_metadata
+from .find_dicoms import find_dicoms
 from .input import (
     extract_roi_meta,
     extract_roi_names,
@@ -7,7 +8,7 @@ from .input import (
     load_seg_dcm,
     rtstruct_reference_uids,
 )
-from .utils import find_dicoms, lookup_tag, similar_tags, tag_exists
+from .utils import lookup_tag, similar_tags, tag_exists
 
 __all__ = [
     "find_dicoms",
